fix: resolve SQLite NotSupportedException in GetTotalBalanceAsync
Modified TransactionService.GetTotalBalanceAsync to cast decimal Balance to double during summation, as SQLite doesn't support Sum on decimal types in EF Core. This follows the established pattern in the codebase for other sum operations.

+++ src/Kopilka.BusinessLogic/TransactionService.cs
@@ -212,9 +212,9 @@ namespace Kopilka.BusinessLogic
         /// <returns>Общий баланс.</returns>
         public async Task<decimal> GetTotalBalanceAsync(int userId)
         {
-            return await _context.Accounts
+            return (decimal)await _context.Accounts
                 .Where(a => a.UserId == userId)
-                .SumAsync(a => a.Balance);
+                .SumAsync(a => (double)a.Balance);
         }
 
         /// <summary>
